Finished.

+++ server/src/service/message/delete-message.service.ts
@@ -1,10 +1,24 @@
-import { Conversation, Message } from "../../model";
+import { Conversation, Message, User } from "../../model";
+import { imageService } from "../image.service";
+import { ApiException } from "../../exception";
 
-export const deleteMessageService = async ( messageId: number, conversationId: number ) => {
+export const deleteMessageService = async ( currentUserId: number, messageId: number, conversationId: number ) => {
 
-   await Message.destroy({ where: { id: messageId } });
+   const [ message, user ] = await Promise.all( [
+      Message.findOne( {
+         where: { conversationId, id: messageId },
+         attributes: [ "content", "isImage" ]
+      } ),
+      User.findByPk( currentUserId )
+   ] );
 
-   return await Conversation.findByPk(conversationId, {
+   if ( !message ) throw new ApiException( "Message does not exist", 400 );
+
+   if ( message?.isImage && user ) await imageService.delete( user.email, message.content );
+
+   await Message.destroy( { where: { id: messageId } } );
+
+   return await Conversation.findByPk( conversationId, {
       include: {
          model: Message,
          as: "lastMessage"
@@ -12,7 +26,7 @@ export const deleteMessageService = async ( messageId: number, conversationId: n
       order: [
          [ { model: Message, as: "lastMessage" }, "id", "DESC" ]
       ]
-   })
-       .then(conversation => conversation?.lastMessage);
+   } )
+       .then( conversation => conversation?.lastMessage );
 
 };
